Test work

+++ delite.php
@@ -1,7 +1,16 @@
 <?php
 session_start();
 
-unlink('file.php');
-$_SESSION['message_no_good'] .= 'Файл удален \n';
-header("Location: ".$_SERVER["HTTP_REFERER"]);
-exit;
+$filename = 'file.php';
+if (file_exists($filename)){
+    unlink('file.php');
+
+    $_SESSION['message_good'] .= 'Файл удален \n';
+    header("Location: ".$_SERVER["HTTP_REFERER"]);
+    exit;
+}
+else {
+    $_SESSION['message_no_good'] .= 'Файла на удаление нет! \n';
+    header("Location: ".$_SERVER["HTTP_REFERER"]);
+    exit;
+}
